spec: Add sample generator code using OpenSSL
Add sample generator code to the specification.  The code is
compile-tested as part of the build.

The use of Pygments for code highlighting requires LyX 2.3 or newer.
The document must also be authorised to run external commands: this
can be done via

  Document -> Settings -> Formats -> Allow running external programs

The CI pipeline installs the latest LyX release (since Ubuntu does not
yet ship version 2.3) and authorises the document by directly creating
the relevant entry in a new ~/.lyx/session file.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>