spec: Use SEQUENCE OF rather than SET OF for encoding preseed values
The preseed values are logically an unordered set rather than a
sequence, but an arbitrary order must be imposed to allow a repeatable
signature to be generated over the DER representation.

Fix by defining the PreseedValues type as a SEQUENCE rather than a
SET, and clarify in the specification wording that the ordering is
arbitrary.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>