android: Add on-device Generator benchmarks.
Generating Contact Identifiers efficiently is important for the
feasibility of the CX model on real devices. This commit introduces some
initial benchmarks for generating a full sequence of 2048 Contact
Identifiers in order to measure this performance.

The benchmarks run on a physical Android device using the Jetpack Benchmark
library https://developer.android.com/jetpack/androidx/releases/benchmark. This
was chosen as it handles many of the complexities involved with physical
device testing and Android benchmarks (read 'JVM JIT compilation') in
general e.g. JVM warmup, measuring performance on the device, outputting
results from the device etc.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>