spec: Refactor SeedReport layout to allow for easier key rotation
Refactor the Seed Report layout to allow the Advertiser to generate a
new Preseed Key Pair for each new Seed Value.  This eliminates the
difficulties of determining when a Preseed Key Pair can safely be
discarded.

Clarify the expected behaviour when a Seed Report is received for a
Seed Value that has already been seen in a previous Seed Report.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>