doc: Provide useful target links for the CI status badges
The GitHub-suggested CI status badges link by default to the SVG image
itself, which is not particularly informative.  There is no simple way
to link to the most recent test log a la Travis CI, but it is possible
to construct a query that will at least show all test logs from the
relevant workflow for the master branch.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>