spec: Document role and operation of peer subscription zones
Some subscription zones will expect to see high volumes of travel to
and from other nearby Subscription Zones.  For example: political or
commercial factors can result in multiple Publishers operating in a
single overlapping geographical area, and frictionless border
arrangements can result in international travel forming part of a
daily commute.  This can result in excessive volumes of notifications
being published to an upstream subscription zone.

The existing protocol allows for this problem to be solved by
configuring additional publications with the "zone" property set to
that of a neighbour rather than to an upstream zone.

Document this existing capability into a formal definition of a peer
subscription zone.

Suggested-by: Luke Barone-Adesi <luke.barone-adesi@unipart.io>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>