python: Allow building against in-tree or externally installed C library
Extend the build_ext command to allow for using either:

- an in-tree copy of the libcx C library, with the autotools-based
  build triggered automatically before attempting to build the Python
  extension, or

- an externally installed copy of the libcx C library, detected via
  pkg-config.

The in-tree copy is used by default if building from a source tree
that includes the C library.  This can be overridden on demand using
the "--extlibcx" option.

Also use the pkg-config infrastructure to link directly against
OpenSSL since the Python extension does call into OpenSSL directly (as
well as indirectly via libcx).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>