spec: Emphasise the capability for revocation of earlier alerts
Modelling carried out for NHSX ("Effective Configurations of a Digital
Contact Tracing App: A report to NHSX") suggests that an important
factor in the effectiveness of any contact tracing approach is the
ability to provide fast alerts of self-reported symptoms combined with
the ability to subsequently revoke the alert if a test proves
negative.

This is already supported by the CX protocol, since an alertLevel will
supersede any earlier reported alertLevel and there is already a
defined alertLevel of no-alert(0).  Emphasise this ability within the
cover page bullet points (replacing the point about immediate
deployability, since there is no more space on the cover page).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>