python: Allow extension to be built from sdist without Cython installed
Avoid an import failure on running setup.py if Cython is not
installed, and allow extension building to rely upon the transparent
support for Cython within setuptools.  The net result is that the
extension may be built from an sdist on a system that does not have
Cython installed (or that has an out of date version of Cython).

If Cython is installed then we use cythonize() explicitly since the
transparent support available from setuptools does not provide a way
to specify language level or to request that an annotation file is
generated.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>