c: Declare test vector files as standard automake sources
Declare the test vector source files (used verbatim from the
specification document) as standard automake sources, rather than as
EXTRA_DIST files that are not dependency-tracked.

This requires removing the "static" declarations and ensuring that
variable names are unique within the test vector files.

Declare the public key .c files as sources and compute the list of
.pem files from this, since automake will not expand $(patsubst) when
generating Makefile rules.

Remove the use of BUILT_SOURCES, since the compilation of the public
key .c files is no longer invisible to automake.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>