c: Include ASN.1 library autogenerated using asn1c
Include an ASN.1 library generated from the CX ASN.1 schema using
asn1c.  This library is built without -Werror as a libtool convenience
library to allow for the quirks of the autogenerated code.

The pristine sources include only a Makefile.am fragment containing
the names of the autogenerated files.  This is required since automake
needs to know the full sources list before generating the build rules.
The Makefile.am fragment will be rebuilt automatically if the ASN.1
schema files are changed; the modifed fragment should then be
committed along with the schema change.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>