spec: Move Signature definition to general section of ASN.1 schema
Similar to the Extension object, the Signature object is now general
and is no longer intrinsically tied to the concept of a Seed Report.
Adjust the ordering of the ASN.1 schema to reflect this.

Since the Signature is currently used only within a Seed Report, leave
the location of its documentation unaltered.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>