kotlin: Add minimal plugin configuration for Android builds
Add a minimal Gradle plugin configuration required for enabling
Android builds.

Almost all existent 'documentation' on this topic seems to suggest
using "buildscript" and "apply" blocks in order to work around what
seems to be a longstanding incompatibility between the Android Gradle
Plugin and the plugin naming scheme expected by Gradle.  This in turn
imposes a limitation that the type-safe accessors are not available
for use within the same build.gradle.kts file, but only within the
build.gradle.kts files of child projects.  This appears in turn to be
the reason why so many codebases have apparently unnecessary child
projects.

Many thanks to StefMa for the article "It's time to ditch the
buildscript block", which provided the basis of this alternative
approach.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>