kotlin: Build OpenSSL as an Android native library
Build OpenSSL using the Android NDK via the externalNativeBuild
mechanism.

The Kotlin Multiplatform Gradle plugin is unfortunately not designed
according to the Unix philosophy of "do one thing and do it well"; it
is a huge and lumbering monolith.  Though it does provide the ability
to compile Android native code via the NDK (using a copy of the
Android NDK that is completely unrelated to the copy used by the
Android Gradle plugin), it provides no sensible way to retrieve the
locations of these downloaded tools.

The older Kotlin Native Gradle plugin did provide clean ways to access
the individual NDK tools.  Unfortunately, this plugin is now
deprecated and has a package namespace collision with the
substantially less friendly Kotlin Multiplatform Gradle plugin.

The upshot is that there is no clean way to exploit Kotlin
Multiplatform's capability to automatically download appropriate
toolchains for each supported target architecture.  Provide instead an
Android-only mechanism using CMake to drive the OpenSSL build process.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>