kotlin: Produce a single shared object
Android doesn't really make use of shared libraries in any sensible
way: there is no real packaging infrastructure and no ability for
multiple apps to link against the same shared library at runtime.

Produce a single .so file using a wrapper that simply defines
requirements for the required JNI_OnLoad and JNI_OnUnload symbols.
This has the advantage that the build system is fully aware of the
final binary (without any file-copying hacks), and so ensures that
problems such as libraries being built for the wrong CPU architecture
will result in a build failure rather than a runtime failure.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>