kotlin: Rewrite Generator to avoid use of JNI
The use of JNI for operations supporting cryptographic keys on Android
would be painful, since it would either be restricted to using only
unprotected software keys, or would require the implementation of an
OpenSSL engine to proxy requests back up to the Java layer.

Switch to using a pure JVM implementation of the Generator
functionality, based on the existing Android implementation.

Rewrite the native implementation to use a more natural pattern of
calls into the libcx C library, since we are no longer constrained to
use a mechanism that keeps the JNI code simple.

Originally-implemented-by: Ben Steadman <steadmanben1@gmail.com>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>